Move SQLite class to data_access folder
